+++ src/types.d.ts
@@ -90,7 +90,7 @@ declare global {
                 xp: { value: number, max: number },
                 level: { value: number }
             };
-            exploration: string[];
+            exploration?: string[];
         };
     }